requirements.txt
# Site
Flask

# Tracer
google-api-python-client
google-auth-httplib2
google-auth-oauthlib